Promote VarCollector to crate-level utility
