+++ ecmascript/transforms/src/modules/util.rs
@@ -4,7 +4,7 @@ use fxhash::{FxBuildHasher, FxHashMap, FxHashSet};
 use inflector::Inflector;
 use std::iter;
 use swc_atoms::JsWord;
-use swc_common::{Mark, Span, SyntaxContext, Visit, VisitWith, DUMMY_SP};
+use swc_common::{Mark, Span, SyntaxContext, DUMMY_SP};
 
 #[derive(Clone, Default)]
 pub(super) struct Scope {
@@ -419,39 +419,6 @@ pub(super) fn make_descriptor(get_expr: Box<Expr>) -> ObjectLit {
         ],
     }
 }
-
-pub(super) struct VarCollector<'a> {
-    pub to: &'a mut Vec<(JsWord, SyntaxContext)>,
-}
-
-impl<'a> Visit<VarDeclarator> for VarCollector<'a> {
-    fn visit(&mut self, node: &VarDeclarator) {
-        node.name.visit_with(self);
-    }
-}
-
-impl<'a> Visit<Ident> for VarCollector<'a> {
-    fn visit(&mut self, i: &Ident) {
-        self.to.push((i.sym.clone(), i.span.ctxt()))
-    }
-}
-
-macro_rules! var_noop {
-    ($T:path) => {
-        impl<'a> Visit<$T> for VarCollector<'a> {
-            fn visit(&mut self, _: &$T) {}
-        }
-    };
-
-    ($T:path, $($rest:tt)*) => {
-        var_noop!($T);
-        var_noop!($($rest)*);
-    };
-}
-
-var_noop!(Expr, ArrowExpr, Function, Constructor);
-
-var_noop!(TsType, TsTypeAnn, TsTypeParam);
 
 /// Private `_exports` ident.
 pub(super) struct Exports(pub Ident);
